backend/requirements.txt

# Backend Requirements (mirrors project top-level requirements)
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Redis
redis==5.0.1
hiredis==2.2.3

# Vector Store
# Use Qdrant as the self-hosted vector DB. The Python client is lightweight
# and connects to the Qdrant REST API exposed by the Docker service.
qdrant-client>=1.5.0,<2

# LangChain Ecosystem & LangGraph
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
langgraph>=0.0.20
openai==1.6.1

# Azure
azure-ai-formrecognizer==3.3.2
azure-storage-blob==12.19.0
azure-identity==1.15.0

# PDF Processing
pdfplumber==0.10.3
pypdf==3.17.1

# Web Scraping & RSS
feedparser==6.0.10
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1

# Email
imap-tools==1.6.0

# Utilities
python-dotenv==1.0.0
# Pin httpx to the 0.27.x series to remain compatible with the OpenAI/Azure client
# and still satisfy chromadb's requirement (chromadb requires httpx>=0.27.0).
httpx==0.27.2
websockets==12.0

# Optional: Gemini (fallback)
# google-generativeai==0.3.1

